Create NPYtoCLEAN_NPY.py
The code that converted the .jpegs to .npy files did not handle greyscale images and incorrectly-sized images. This script handles those explicitly so that now all images should be the same size and have the same number of dimensions. The directory that this script accesses to make the clean .npy files consists of intermediate files created during JPEGtoNPY.py .